+++ assets/js/app.ts
@@ -1,4 +1,4 @@
-import {Item, View} from 'vega';
+import * as vega from 'vega';
 import vegaEmbed from 'vega-embed';
 
 declare global {
@@ -16,9 +16,11 @@ interface SelectOptions {
 
 const viewsById = {};
 const heightsById = {};
+const chartRefreshRate = 5000; // milliseconds
+const chartDomainBuffer = 60000; // milliseconds
 
 const onSelectTag = (
-  view: View,
+  view: vega.View,
   selectedItem: HTMLElement,
   legendItems: HTMLElement[],
   options: SelectOptions
@@ -129,8 +131,18 @@ const renderLegend = (id: string, viewUnknown: unknown) => {
   if (!legend || !runtime.scales.color) return;
 
   legend.classList.remove('hidden');
+  const itemsCount = Array.from(
+    legend.querySelectorAll('[telemetry-component="LegendItem"]')
+  ).length;
+
   const colors = view.scale('color').range();
   const categories = view.scale('color').domain();
+
+  if (itemsCount !== 0 && itemsCount === categories.length) return false;
+
+  const selectedValues = Array.from(
+    legend.querySelectorAll('[data-selected="true"]')
+  ).map((item) => item.dataset.value);
 
   const content = categories
     .sort((a, b) => (parseInt(a) || a) - (parseInt(b) || b))
@@ -141,9 +153,11 @@ const renderLegend = (id: string, viewUnknown: unknown) => {
 
       if (!category) return;
 
+      const categoryValue = btoa(category);
+
       return `<span
             telemetry-component="LegendItem"
-            data-value="${btoa(category)}"
+            data-value="${categoryValue}"
             class="truncate max-w-full cursor-pointer hover:bg-neutral-50 hover:dark:bg-neutral-800 shrink-0 px-2 py-1 inline-block rounded-sm border-[1px] border-neutral-200 dark:border-neutral-800"
           >
             <span class="inline-block rounded-full h-[10px] w-[10px]" style="background: ${color};"></span>
@@ -154,6 +168,18 @@ const renderLegend = (id: string, viewUnknown: unknown) => {
     .join('\n');
 
   legend.innerHTML = content;
+
+  const legendItems = Array.from(
+    legend.querySelectorAll('[telemetry-component="LegendItem"]')
+  ) as HTMLElement[];
+
+  legendItems.forEach((item) => {
+    if (selectedValues.includes(item.dataset.value)) {
+      onSelectTag(view, item, legendItems, {multi: true});
+    }
+  });
+
+  return true;
 };
 
 const emptySource = (source) => {
@@ -214,25 +240,60 @@ window.drawChart = async (id, spec) => {
       editor: false
     }
   });
-  const element = document.querySelector(id) as HTMLElement;
 
-  if (emptySource(view.data('source'))) {
-    const title = document.querySelector(id + '-title') as HTMLElement;
-    title.classList.remove('hidden');
+  if (spec.data.url) {
+    setInterval(async () => {
+      const response = await fetch(spec.data.url);
+      const data = await response.json();
 
-    const empty = document.querySelector(id + '-empty') as HTMLElement;
-    empty.classList.remove('hidden');
-    element.classList.add('hidden');
-    element.classList.remove('vega-embed');
-  } else {
-    renderLegend(id, view);
-    bindLegend(element, id, view);
+      const now = Number(new Date());
+      const [currentDomainFrom, currentDomainTo] = view.signal('date_domain');
+      let changeset = vega
+        .changeset()
+        .remove(() => true)
+        .insert(data);
+      view.change('source', changeset);
+      view.signal('date_domain', [
+        now - (currentDomainTo - currentDomainFrom - chartDomainBuffer),
+        now + chartDomainBuffer
+      ]);
+
+      view.run();
+      window.dispatchEvent(new Event('resize'));
+
+      setSideElements(view, id);
+    }, chartRefreshRate);
   }
+
+  setSideElements(view, id);
 
   const loading = document.querySelector(id + '-loading') as HTMLElement;
   loading.classList.add('hidden');
 
   viewsById[id] = view;
+};
+
+const setSideElements = (view, id) => {
+  const title = document.querySelector(id + '-title') as HTMLElement;
+  const element = document.querySelector(id) as HTMLElement;
+  const legend = document.querySelector(id + '-legend') as HTMLElement;
+  const empty = document.querySelector(id + '-empty') as HTMLElement;
+
+  if (emptySource(view.data('source'))) {
+    title.classList.remove('hidden');
+    empty.classList.remove('hidden');
+    element.classList.add('hidden');
+    element.classList.remove('vega-embed');
+    legend.classList.add('hidden');
+  } else {
+    title.classList.add('hidden');
+    empty.classList.add('hidden');
+    element.classList.remove('hidden');
+    element.classList.add('vega-embed');
+
+    const renderedLegend = renderLegend(id, view);
+    if (renderedLegend) bindLegend(element, id, view);
+  }
 };
 
 document.addEventListener('DOMContentLoaded', () => {