+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `TelemetryUI.Web.Component.to_image` implementation of VegaLite components has been extracted to a standalone library to remove hard dependency on rustler precompiled and vix.
 
+### Features
+
+- Interval refresh for live charts update :tada:
+
 ## v3.0.0 (2023-04-26)
 
 ### BREAKING CHANGES